EndBlock fix, enhanced fill in btape, ...
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@240 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/dird/restore.c
@@ -330,7 +330,7 @@ Termination:            %s\n\n"),
 	jcr->client->hdr.name,
 	sdt,
 	edt,
-	edit_uint64_with_commas(jcr->jr.JobFiles, ec1),
+	edit_uint64_with_commas((uint64_t)jcr->jr.JobFiles, ec1),
 	edit_uint64_with_commas(jcr->jr.JobBytes, ec2),
 	(float)kbps,
 	term_msg);
